feat(slides): add student slides nav + fix Vite 6 resolve conditions
Add STUDENT role to slides nav in routes.js so students see the
Slides link in their sidebar. Extract root loader to root.server.js
following React Router conventions.

Fix @prisma/client resolution error caused by @react-router/dev
resolving its vite peer dep to root Vite 5 (which lacks
defaultClientConditions), leaving client resolve conditions empty.
Explicitly set Vite 6 default conditions for both client and SSR.
