+++ apps/slides/vite.config.js
@@ -1,5 +1,5 @@
 import { reactRouter } from '@react-router/dev/vite';
-import { defineConfig } from 'vite';
+import { defineConfig, defaultClientConditions, defaultServerConditions } from 'vite';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import tailwindcss from '@tailwindcss/vite';
 
@@ -7,8 +7,15 @@ export default () => {
   return defineConfig({
     ssr: {
       noExternal: ['reveal.js'],
+      resolve: {
+        conditions: ['development', ...defaultServerConditions],
+      },
     },
     resolve: {
+      // Vite 6 default client conditions must be set explicitly because
+      // @react-router/dev resolves its vite peer dep to root node_modules (Vite 5)
+      // which doesn't export defaultClientConditions, leaving conditions empty.
+      conditions: ['development', ...defaultClientConditions],
       alias: {
         '.prisma/client/index-browser': '../../node_modules/.prisma/client/index-browser.js',
       },
